+++ pkgs/test_core/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## 0.4.14-dev
 
+* Update the github reporter to output the platform in the test names when
+  multiple platforms are used.
 * Fix `spawnHybridUri` support for `package:` uris.
 
 ## 0.4.13